Rename env/ to platform/
