+++ platform/platform_test.go
@@ -1,4 +1,4 @@
-package env
+package platform
 
 import (
 	"testing"
